Metodos separados a sus clases
