+++ Biblioteca/src/Almacen/prueba3.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?><Tests xmlns="http://www.adatum.com"><Test><Name>Convert number to string</Name><CommandLine>Examp1.EXE</CommandLine><Input>1</Input><Output>One</Output></Test><Test><Name>Find succeeding characters</Name><CommandLine>Examp2.EXE</CommandLine><Input>abc</Input><Output>def</Output></Test></Tests>
+<?xml version="1.0" encoding="UTF-8" standalone="no"?><Tests xmlns="http://www.adatum.com"><Test><Name>Convert number to string</Name><CommandLine>Examp1.EXE</CommandLine><Input>1</Input><Output>One</Output></Test><Test><Name>Find succeeding characters</Name><CommandLine>Examp2.EXE</CommandLine><Input>abc</Input><Output>def</Output></Test><Test><Name>Convert number to string</Name><CommandLine>Examp1.EXE</CommandLine><Input>1</Input><Output>One</Output></Test><Test><Name>Convert number to string</Name><CommandLine>Examp1.EXE</CommandLine><Input>1</Input><Output>One</Output></Test><Test><Name>Convert number to string</Name><CommandLine>Examp1.EXE</CommandLine><Input>1</Input><Output>One</Output></Test></Tests>
