--- Biblioteca/src/Errores/Logging.txt.1
@@ -1,2 +1,0 @@
-oct. 23, 2020 8:30:31 A. M. reto.Lector leer
-WARNING: Metodo leer fallo, Seleccionado tipo de archivo no compatible
